Modified road-runner to use the new dockerops package
